\name{toy5}
\alias{toy5}
\docType{data}
\title{
Toy data set with 5 clusters
}
\description{
This artificial data set contains 5 two dimensional gaussian clusters.
}
\usage{data(toy5)}
\format{
toy5 is a matrix with 50 cases (rows) and 2 variables (cols).
}
\examples{
data(toy5)
plot(toy5)
}
\keyword{datasets}
