\name{toy9}
\alias{toy9}
\docType{data}
\title{
Toy data set with 9 clusters
}
\description{
This artificial data set contains 9 two dimensional gaussian clusters.
}
\usage{data(toy5)}
\format{
toy9 is a matrix with 90 cases (rows) and 2 variables (cols).
}
\examples{
data(toy9)
plot(toy9)
}
\keyword{datasets}
