\name{toy5}
\docType{data}
\alias{toy5}
\title{Toy data set for 5 clusters}
\description{
This two-dimensional toy data set illustrates five clusters for cluster
algorithms. Each cluster contains 10 points.}
\usage{toy5}
\format{A matrix containing 50 points in 2 dimensions.}
\source{Each cluster was build by a two-dimensional normal distribution.}
%\references{McNeil, D. R. (1977) \emph{Interactive Data Analysis}.New York: Wiley.}
\keyword{datasets}